Apps/AgriAssist/requirements.txt

# -r requirements.txt

streamlit
requests
Pillow
python-dotenv
langchain
langchain-openai
openai